refactor(runs): decompose pause.test.ts below 500-line limit (warren-0e98)
Extract shared file-local helpers into src/runs/pause.test-helpers.ts
and split describe("tickPauseDetector") into src/runs/pause.tick.test.ts.
Remove the grandfathered file-size budget entry. No test bodies changed;
23 pass / 0 fail.

Closes warren-0e98, warren-b7fa
